fix: ai: fix FREQ trailing garbage, PATCHWORK split-patch algorithm, DCT Block overflow
- FREQ: strip non-printable trailing bytes from BCH padding artifacts
- PATCHWORK: rewrite with split-patch approach (left/right half of same region)
  uses adjacent pixels which have near-zero baseline difference, strength=60
- DCT Block: fix uint8 overflow by explicit float casting
- SPREAD: default strength 6->15 for stronger signal

+++ python-core/core/watermarks/dct_block.py
@@ -49,12 +49,12 @@ class DctBlockWatermark(BaseWatermark):
                     break
                 block = arr[by*block_size:(by+1)*block_size, bx*block_size:(bx+1)*block_size]
                 dct_block = self._dct_2d(block)
-                mid = dct_block[4, 1]  # mid-frequency coefficient
-                q = round(mid / strength)
+                mid = float(dct_block[4, 1])  # mid-frequency coefficient
+                q = int(round(mid / strength))
                 if bi < len(bits):
-                    target = (q // 2) * 2 * strength + (bits[bi] * strength)
+                    target = float((q // 2) * 2 * strength + (int(bits[bi]) * strength))
                 else:
-                    target = q * strength
+                    target = float(q * strength)
                 dct_block[4, 1] = target
                 modified[by*block_size:(by+1)*block_size, bx*block_size:(bx+1)*block_size] = self._idct_2d(dct_block)
                 bi += 1
